+++ cuppa/reports/__init__.py
@@ -1,0 +1,8 @@
+#          Copyright Jamie Allsop 2026-2026
+# Distributed under the Boost Software License, Version 1.0.
+#    (See accompanying file LICENSE_1_0.txt or copy at
+#          http://www.boost.org/LICENSE_1_0.txt)
+
+#-------------------------------------------------------------------------------
+#   Cuppa CLI report manifest (interim cleanup until artefact_roots lands)
+#-------------------------------------------------------------------------------